We have a new platform now, its openbsd.
There are some minor changes for the scripts. We have the flavour now as a number, and its
diffrent for the unixes now.
for the cygwin and mingw a 1 is fine.
For the openbsd we need also now pthread specific data instead of thread local storage.
They simply don't have its and so the new clang is still out of option here.