We have fixed some linker messages for openbsd.
They have made clear they take security serious.
So they make the linker noisy if you try something like strcpy or strcat or sprintf ...

Which is a good thing if you work with offshore contracters....

But as always - my code did most of that things WITH the lenght in mind ....

There are some exceptions, and they are in clear violation of the max lenght of a path on the systems fielsystem...

OK.

So we have now the flavour 3 and 4 back in code again.

The mentioned functions of bsd produce in realety a FALSE sense of security.

Example:

strcpy(dest, src);

Thats dangerous if you NOT know how big dest is - if it is a buffer after all - and if src could be a c string larger than dest - if dest is a buffer after all ....

So change it into
strlcpy(dest,src, lenght);

make length a variable and a simple array off by one writes, let say  - 23456453897897 - in it on the stack ....
then  your so called strlcpy is worth not even the paperwork i have with it ....

or:

the guy tried to make it GOOD by useing a define constant ...

strlcpy(des, src, MY_DEST_LEN);

Well, thats good if everyone agrees about it - but i had in the past a bunch of projects that had -
hand copied and then hand crafted headers, and at some point ...
you get it.

A define is not a real solution.

How about this one ?

strlcpy(dest, src, sizeof(dest));

Looks good - but if dest is a pointer from an alloc - whichever and how safe leave i to you - well, then the thing is simply WRONG....

So the C standard and the Linux community seems to abandon the strlcpy and the siblings. only snprintf is accepted....

So i made the snprintf and have now twice the code to maintaine ... thanx, openbsd.

And no additional security against misuse ...