We have a new platform now, netbsd.
The scripts have a new flavour 5 to handle.
The includes have the new defiens for the thread id on netbsd and a new helper
for the safer copy functions.

For the code its a new non inline and the rest the same as the gcc linux.
Only some includes differ as it is on bsd.

For the gcc 6.2 i got a new check program - it has no threads.h, but the define
is set as if it had one...