New slave sruct replace link list former via thread local.
Now we have a slave struct.
So we can now handle the kill of slave better.
The list is still intact, no thread local here.
So after the kill you can iterate it, find the tid and so the slave adress for remove slave local.

This means also new functions for get tid via slave and turn off for a slave.

Beside minor fixes.