We have updates swig according to the perl changes.
We still do not deliver the poke, the peek and the tests.

This is because we simply assume the tcl will only be a test drive for swig.

If you insist i can add tcl to the layers and we have a different swig then for tcl and with tests.
simply sent me a postcard ...