We have a new helper : generate the prealloc buffers file
We have a new perl helper.
It generates the impls/..prealloc buffer file.
So we have no longer to adjust by hand, only to set the new values and hit the makeall button.
The thing only writes a proposal file - you have to do the final copy by yourself then.
This is the only safety, but its a good one.