We have a new converter. Its convert to postgres db
We have aschema file for a database.
You decide what it is.

Then we have the new converter that takes files and convert tehm into content for the db.

Later on we will integrate this into a new reader.

The stuff is db bependent, so we start with explicit postgres.
later on mysql and oracle ...
rest up to you.