We have decoupled the db part and the main logic and the driver starts
We have now the main in separate files for the db stuff.
so we can link against it and get the name of the program

its a simple call to the db indemendent logics main.

we have also the db specific now only one time in in a separate file.

we have only to link and to name and then get a new db done ...