We have negated the database part because of the dependency of oralce
So if we want the database we have to decide to bring it in.
And we have to make the env proper with oralce flags in the CFLAGS
we have the oralce things in place in development (my favorite db after all)
and in unsupported the needed helper scriptings.
If you want the database but no oralce you have to remove the oracle and ora files from dbs and the shm file should be empty.