New handling for disposed buffers.
We dispose mean we set our ptr to null.
So in write we honor also null ptrs and if no buffer is nonnull we discart.
The flush is also done the dispatch only if it is needed. so no wait for
empty buffers now. only buffers with content or dispose are dispatched.